requirements.txt
numpy==2.4.1
pandas == 3.0.0
python-dotenv==1.2.1